+++ src/core/input.spec.ts
@@ -100,18 +100,18 @@ describe('Input functions and helpers', () => {
   })
 
 
-  describe('Generic action self caller', () => {
-    let actionData = []
+  describe('Generic action self call', () => {
+    let actionData
     let root: Component<any> = {
       name: 'MyComp',
       state: 0,
       inputs: ctx => ({
-        action: ([name, data]) => {
-          actionData = [name, data]
-        },
       }),
       actions: {
-        Inc: s => s + 1,
+        Name: data => s => {
+          actionData = data
+          return s
+        },
       },
       interfaces: {},
     }
@@ -122,9 +122,9 @@ describe('Input functions and helpers', () => {
 
     it('should call input with the action name and data', () => {
       toAct(app.ctx.components.MyComp.ctx)('Name', 'data1')
-      expect(actionData).toEqual(['Name', 'data1'])
+      expect(actionData).toEqual('data1')
       toAct(app.ctx.components.MyComp.ctx)('Name', 'data2', false, true)
-      expect(actionData).toEqual(['Name', 'data2'])
+      expect(actionData).toEqual('data2')
     })
 
   })